# employee-services/openapi.yaml
openapi: 3.1.0
info:
  title: Internal Transfer API (employee-services)
  version: "1.0.0"
  description: >
    HTTP surface for the internal-transfer module. Covers the employee request API
    (API01-API07), the approval-chain API, and the downstream stage-completion webhook.
    The notifications service exposes no HTTP route; its portal -> notification-service
    payload is the contract in docs/contracts/notification-dispatch.v1.md and is referenced
    from the approval and request events, not served from here.
    Added for Gate 2 finding G2-F06.
servers:
  - url: http://localhost:3001
    description: Local demo
...
paths:
...
  /api/v1/internal-transfers/approvals:
    get:
      tags: [approval]
      summary: Approver inbox
      description: Rate limit 300/hour per principal. Shows only the caller's IN_PROGRESS stages (manager = assigned party; HR = any HR_VALIDATION). No reason text.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1 } }
        - { name: size, in: query, required: false, schema: { type: integer, enum: [10, 25, 50] } }
      responses:
        "200":
          description: Paginated inbox
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/InboxItem" }
                  page: { type: integer }
                  size: { type: integer }
                  totalItems: { type: integer }
                  totalPages: { type: integer }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/internal-transfers/{requestId}/approval:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      tags: [approval]
      summary: Approval detail
      description: Rate limit 300/hour per principal. Reason is included only for HR. A caller who cannot see the request gets 404.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Approval detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalDetail" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/RequestNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/internal-transfers/{requestId}/stages/{stageCode}/decision:
    parameters:
      - $ref: "#/components/parameters/RequestId"
      - { name: stageCode, in: path, required: true, schema: { type: string, enum: [MANAGER_RELEASE, MANAGER_ACCEPT, HR_VALIDATION] } }
    post:
      tags: [approval]
      summary: Record a stage decision
      description: >
        Rate limit 30/hour per principal. Idempotency-Key and If-Match required. HR approve
        needs confirmedEffectiveDate; the other stages must not carry one. Only the current
        IN_PROGRESS stage can be decided.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [APPROVE, REJECT] }
                confirmedEffectiveDate: { type: string, format: date, nullable: true }
      responses:
        "200":
          description: Decision recorded; returns the updated approval detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalDetail" }
        "400":
          description: precondition-required (no If-Match) or idempotency-key-required
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/RequestNotFound" }
        "409":
          description: version-conflict, invalid-state-transition, idempotency-key-conflict, or assignee-unresolved
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/internal-transfers/webhooks/stage-completion:
    post:
      tags: [webhook]
      summary: Downstream stage-completion report (API01 downstream)
      description: >
        Rate limit 600/hour per configured source. Authenticated by HMAC signature over the
        canonical string, not by an employee token; a bearer token with no signature is 401.
        eventId is the idempotency key: a byte-identical replay returns 200; the same eventId
        with a different body is 409. Only the current IN_PROGRESS stage can be reported.
      security: [{ webhookSignature: [] }]
      parameters:
        - { name: X-Portal-Webhook-Key-Id, in: header, required: true, schema: { type: string } }
        - { name: X-Portal-Webhook-Timestamp, in: header, required: true, schema: { type: string, format: date-time }, description: "ISO, within +/-300s of portal time" }
        - { name: X-Portal-Webhook-Signature, in: header, required: true, schema: { type: string }, description: "sha256=<hex> HMAC of the canonical string" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StageReport" }
      responses:
        "200":
          description: Accepted (including a byte-identical replay)
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: { type: boolean }
        "401":
          description: Missing, invalid, replayed (timestamp outside +/-300s) or unknown-key signature; nothing logged from the body
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: version-conflict (lost CAS) or idempotency-key-conflict (same eventId, different body)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Portal session bearer token. Subject is the identity; roles are resolved server side (G2-F02).
    webhookSignature:
      type: apiKey
      in: header
      name: X-Portal-Webhook-Signature
      description: HMAC-SHA256 over the five-line canonical string using the source's signing secret.
  parameters:
    RequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
      description: Quoted aggregate version for compare-and-swap (OWN-11)
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
  responses:
    Unauthenticated:
      description: Missing or invalid credentials; no state change
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RequestNotFound:
      description: Not found, or not visible to the caller (ownership returns 404, not 403)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationFailed:
      description: Request body failed validation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RateLimited:
      description: Rate limit exceeded; Retry-After header present, no state change (AC17)
      headers:
        Retry-After:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807-style problem. 'type' carries the machine-readable code.
      properties:
        type: { type: string, example: rate-limited }
        title: { type: string }
        status: { type: integer }
        violations:
          type: array
          items:
            type: object
            properties:
              ruleId: { type: string }
              field: { type: string }
...
    InboxItem:
      type: object
      properties:
        requestId: { type: string }
        referenceNo: { type: string }
        stageCode: { type: string }
        assignedRole: { type: string }
        requestStatus: { type: string }
        targetPositionTitle: { type: string }
        requestedEffectiveDate: { type: string, format: date, nullable: true }
        waitingSince: { type: string, nullable: true }
    ApprovalDetail:
      type: object
      properties:
        requestId: { type: string }
        referenceNo: { type: string }
        status: { type: string }
        version: { type: integer }
        confirmedEffectiveDate: { type: string, format: date, nullable: true }
        reason: { type: string, description: "Present only for HR" }
        stages:
          type: array
          items:
            type: object
            properties:
              stageCode: { type: string }
              status: { type: string }
              assignedRole: { type: string }
    StageReport:
      type: object
      required: [eventId, requestId, stageCode, reportType, outcome]
      properties:
        eventId: { type: string, format: uuid, description: "Idempotency key (BR11)" }
        requestId: { type: string, format: uuid }
        stageCode: { type: string, enum: [ORG_DATA_UPDATE, PAYROLL_UPDATE, IT_ACCESS, FACILITIES] }
        reportType: { type: string, enum: [FULFILMENT, COMPENSATION] }
        outcome: { type: string, enum: [SUCCESS, FAILED] }
        failureCode: { type: string, nullable: true }
        expectedVersion: { type: integer, nullable: true }